+++ test/combinatorics.jl
@@ -37,11 +37,11 @@ end
     @test !isperm(p)
 
     a = randcycle(10)
-    @test ipermute!(permute!([1:10;], a),a) == [1:10;]
+    @test invpermute!(permute!([1:10;], a),a) == [1:10;]
 
     # PR 12785
     let ai = 2:-1:1
-        @test ipermute!(permute!([1, 2], ai), ai) == [1, 2]
+        @test invpermute!(permute!([1, 2], ai), ai) == [1, 2]
     end
 end